Syncing drawings with fullscreen chart
